Functor
This adds an interface that stands in for the type class of Functors.
Unfortunately I haven't yet found a way to wire it to specific
implementations like that in maybe. In an ideal world there would be
some way to say that the package maybe has an implementation of the
Functor interface. Or some type defined in package maybe implements it.

This adds an implementation of Fmap and Freplace for the maybe package.

This also replaces the type switch usage with an explicit type if. The
problems with the switch are that 1) it doesn't realize all cases are
matched requiring a panic and 2) the compiler doesn't infer the matched
type like it usually would requiring yet another type assertion.
Removing it makes the code a little less verbose.