data/either: Don't walk left values
It seemed reasonable that you might want to walk the left value if it
exists, but the expectation seems to be to only walk the left value.
This presents itself in cases like this example from foldable:

```haskell
Prelude Data.Monoid Data.Foldable> concat (Left 42)
[]
```

Where if you change Left to Right you get a type error:

```
Prelude Data.Monoid Data.Foldable> concat (Right 42)

<interactive>:9:1: error:
    • Non type-variable argument in the constraint: Num [a]
      (Use FlexibleContexts to permit this)
    • When checking the inferred type
        it :: forall a. Num [a] => [a]
```

Which can be fixed by changing it to a list:

```
Prelude Data.Monoid Data.Foldable> concat (Right [42])
[42]
```

But still the same on Left results in an empty list:

```
Prelude Data.Monoid Data.Foldable> concat (Left [42])
[]
```

We may need different implementations of this later that walk both
sides. Maybe the bifunctors? There may be many cases where the desire is
only to return the "right most" result. For now, we'll just walk the
right side and figure the rest out when we get to it.