data/list: list transformations
There are a number of "old" list functions that are specialized to
arrays (rather than a foldable). This implements some of them.

Note: I'm trying something here with ~[]A to make these functions type
check with both a list.List[X] and []X. It seems to mostly work, but
whether it is sufficient isn't clear (to me) yet.